move fixtures into separate files
